[MSP430] Add support for the `rpt` assembler directive
The `rpt` directive is used to set the repetition count of MSP430X
extended instructions that have an extension word. This patch adds
support for using the `rpt` directive with RxxX shifts, so shift counts
of up to 16 can be performed using a single instruction.

The `rpt` directive must appear before the corresponding MSP430X
instruction, either on the same line, or the line before. For example:
  rpt #9 { rrux r12
and
  rpt #9
  rrux r12
are both supported.

Differential Revision: https://reviews.llvm.org/D110724